requirements.txt
appdirs==1.4.4
asgiref==3.5.1
attrs==21.4.0
beautifulsoup4==4.11.1
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-ckeditor==6.4.2
django-hCaptcha==0.2.0
django-js-asset==2.0.0
idna==3.3
lxml
model-bakery==1.7.0
python-coinmarketcap==0.3
python-decouple==3.6
ratelimit==2.2.1
requests==2.27.1
requests-cache==0.9.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
url-normalize==1.4.3
urllib3==1.26.9
